requirements.txt
Django==5.1.2
Faker==30.3.0
whitenoise==6.7.0
djangorestframework==3.15.2
